the fix

+++ killtester.java
@@ -3,6 +3,6 @@ import org.junit.*;
 
 public abstract class killtester {
     public void test() {
-        assertEquals(killdemo.killme(), 1);
+        assertEquals((int) (killdemo.killme()), 1);
     }
 }
